Filtering Step 3/6 - HWE
The purpose of this script is to identify loci out of HWE using the exact permutation test. Additionally, this will run the HWE test for multiple 'groups' if you have a larger data set. We then used Fisher's combination of probability test to get one p-value per loci, converted that p-value to a q-value, and filtered our VCF file for loci significantly in HWE.