Filtering Step 6/6 - Linkage
The purpose of this script is to use the genetics package to calculate LD on each chromosome separately. Then, take a directory of vcf files (split by chromosome), calculate LD within each chromosome, and save the output as a heatmap. Finally, plot LD decay plot and summary statistics per chromosome.